+++ client/src/components/Ticketing/cart/Cartmain.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import Cart from './Cart';
 import Navbar from '../mainpage/Navbar';
-import Footer from '../mainpage/footer';
+import Footer from '../mainpage/Footer';
 
 /**
  * @param {Function} Cartmain
