+++ client/src/components/Ticketing/checkout/Checkoutmain.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import CheckoutPage from './CheckoutPage';
 import Navbar from '../mainpage/Navbar';
-import Footer from '../mainpage/footer';
+import Footer from '../mainpage/Footer';
 
 /**
  * Hosts checkout page
