+++ client/src/components/Ticketing/mainpage/EventCard.tsx
