+++ client/src/components/Ticketing/mainpage/Footer.tsx
