+++ client/src/components/Ticketing/mainpage/Hero.tsx
@@ -1,5 +1,5 @@
 import React, {ReactElement, useEffect, useRef} from 'react';
-import {ListComponent} from './eventcard';
+import {ListComponent} from './EventCard';
 import {useAppSelector, useAppDispatch} from '../app/hooks';
 import {fetchTicketingData} from '../ticketingmanager/ticketingSlice';
